requirements.txt
# Core
pandas==2.1.4
numpy==1.26.3
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pytz==2024.1
python-dotenv==1.0.0
openpyxl==3.1.5
xlrd >= 2.0.1
molmass==2025.11.11
chempy
# Geospatial support
geoalchemy2==0.14.3
shapely==2.0.2

# Data Quality
pandera==0.18.0

# Config
pyyaml==6.0.1
python-dotenv==1.0.0

# Utilities
loguru==0.7.2

# Interactive shells
ipython==8.20.0
ipdb==0.13.13

# Notebooks
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.0

# Geospatial visualization (optional)
folium==0.15.1
geopandas==0.14.1

# Testing
pytest==7.4.4
